requirements.txt
aiofiles==0.6.0
anyio==3.0.0
asyncclick==7.1.2.3
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
opencv-python==3.4.13.47
playsound==1.2.2
priority==1.3.0
pycryptodome==3.10.1
pyglet==1.5.16
pytapo==2.0
python-kasa==0.4.0.dev2
pyttsx3==2.90
Quart==0.14.1
requests==2.25.1
rpi==0.0.1
RPi.GPIO==0.7.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
selenium==3.141.0
sniffio==1.2.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
vext==0.7.6
vext.gi==0.7.4
Werkzeug==1.0.1
wsproto==1.0.0
youtube-dl==2021.4.17
zipp==3.4.1
